+++ lib/td_dd/cache/system_loader.ex
@@ -16,15 +16,29 @@ defmodule TdDd.Cache.SystemLoader do
     GenServer.start_link(__MODULE__, config, name: __MODULE__)
   end
 
+  def refresh(system_ids) when is_list(system_ids) do
+    GenServer.call(__MODULE__, {:refresh, system_ids})
+  end
+
+  def refresh(system_id) do
+    refresh([system_id])
+  end
+
+  def delete(system_id) do
+    GenServer.call(__MODULE__, {:delete, system_id})
+  end
+
   ## GenServer callbacks
 
   @impl true
   def init(state) do
     name = String.replace_prefix("#{__MODULE__}", "Elixir.", "")
     Logger.info("Running #{name}")
+
     unless Application.get_env(:td_dd, :env) == :test do
       Process.send_after(self(), :load_cache, 0)
     end
+
     {:ok, state}
   end
 
@@ -33,6 +47,24 @@ defmodule TdDd.Cache.SystemLoader do
     load_all_systems()
 
     {:noreply, state}
+  end
+
+  @impl GenServer
+  def handle_call({:refresh, ids}, _from, state) do
+    reply =
+      ids
+      |> Enum.map(&Systems.get_system/1)
+      |> Enum.filter(&(elem(&1, 0) == :ok))
+      |> Enum.map(&elem(&1, 1))
+      |> load_system_data()
+
+    {:reply, reply, state}
+  end
+
+  @impl GenServer
+  def handle_call({:delete, id}, _from, state) do
+    reply = SystemCache.delete(id)
+    {:reply, reply, state}
   end
 
   ## Private functions