+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-2485]:
+  - Enrich template fields from cache
+  - Mappings for system type of templates
+
 ### Changed
 
 - [TD-3058] Database connection timeout now can be configured using the
   environment variable `DB_TIMEOUT_MILLIS`
-
+  
 ## [4.5.0] 2020-10-05
 
 ### Added